+++ docs/change_log.rst
@@ -49,6 +49,7 @@ The following are only available within Stem's `git repository
  * **Controller**
 
   * Socket based control connections often raised BrokenPipeError when closed
+  * Added :func:`~stem.control.Controller.add_hidden_service_auth`, :func:`~stem.control.Controller.remove_hidden_service_auth`, and :func:`~stem.control.Controller.list_hidden_service_auth` to the :class:`~stem.control.Controller`
 
  * **Descriptors**